# frontend performance: stream to recent topics renarrows