@timabbott 2021-11-04
This is great.  I've never been thrilled with using CSS selectors to check which views are visible, precisely because it can cause subtle performance problems like this.  It is probably worth doing an audit for similar systems in the codebase and replacing those as well.

@amanagr FYI.