@timabbott 2021-11-04
I posted comments on the first commit, which I expect will probably be mergeable after the changes detailed above.  The second commit seems like the sort of thing we should think about other possible solutions for, since as you say it is not pretty and the complexity involved is likely to lead to other bugs.