@timabbott 2021-11-04
BTW another thing that we can potentially consider is having the "Recent topics" view maintain itself rendered in the DOM, like we do for "All messages", so that rendering work happens in the background as new messages arrive (etc.) and switching views to/from it can be instant.