@YashRE42 2021-11-04
Here's a Zip of the profiles that I recorded, which can just be loaded into the profiler for inspection: 
[profiles.zip](https://github.com/zulip/zulip/files/7478444/profiles.zip). I forgot to upload these while formatting the the PR desc.
(I'm heading off for today, will reply tomorrow.)