@timabbott 2021-11-09
I merged the first two commits after changing the API slightly for the new `set_visible` function to accept a boolean, since that seems cleaner.  

I think we should do whatever we choose to do with the last commit in its own PR; it's complicated enough on its own without having other commits in the discussion that are already merged. 

Huge thanks for tracking this down @YashRE42!