@YashRE42 2021-11-14
So, I'd initially been hoping that we could merge the first commit (and its prep commit) and then I could close this or revert to a draft so that the context could all be there. I was planning to make another PR with the last commit and link here...
...but it seems it isn't necessary!
Looking at some profiles on main, I realized our reflow has already been pushed back to upstream simplebar:

![](https://user-images.githubusercontent.com/33805964/141675084-d426cf83-8ba8-4968-b775-e5b5abae5839.png)
[main, a few days later.zip](https://github.com/zulip/zulip/files/7533486/main.a.few.days.later.zip)
(please note that numeric values from the above chrome profile and screenshot are no longer comparable with those recorded earlier in this PR, since my system has had a performance upgrade.)

This is great!
I think it's a side effect of fixing rt to set focus on the correct topic (as part of #18521).
fyi: @timabbott @amanagr.