@timabbott 2021-11-15
Yeah, I had a theory that we'd want to redo this profiling after #18521; glad that the conclusion was that it'd accidentally fixed the other bug here :).